Return errors when fetching rows via Result
This isn't so much of a problem with stored results, but when using
results, it was possible to get disconnected by the server and have it
be entirely indistinguishable from just reaching the end of the table.

This patch adds os.Error return values to FetchMap and FetchRow, which
allow the user to see any errors thus encountered.